requirements.txt
numpy
pandas
scikit-learn
matplotlib
# El sucesor moderno de tmap
umap-learn
# Seaborn para gráficos más bonitos (opcional pero recomendado)
seaborn